move flask stuff to server.py so it can be imported with "from Firefly.server import spawnFireflyServer" etc...
